Indicator Library and EA
Github updated to include new indicator library and EA. This library will store all required functions to set up indicator handles and get indicator signals. This will be used in the Multi-Symbol Indicator EA to process trading signals. The user is able to specify the indicators they wish to use in the EA within the inputs of the EA strategy in metatrader.

Full video tutorial is available on my YouTube channel.